Move model, datasource and project into *Details.js
Removes item_pages.js
